Separate backend routes & supports Cloudflare Workers deployment
