+++ apps/backend/package.json
@@ -5,10 +5,13 @@
   "type": "module",
   "scripts": {
     "build": "tsdown",
+    "build:worker": "tsdown --entry src/worker.ts --outdir dist",
     "check-types": "tsc -b",
     "compile": "bun build --compile --minify --sourcemap --bytecode ./src/index.ts --outfile server",
     "dev": "tsx watch src/index.ts",
+    "dev:worker": "wrangler dev",
     "start": "node dist/index.js",
+    "deploy": "wrangler deploy",
     "test": "vitest",
     "test:run": "vitest run",
     "test:ui": "vitest --ui",
@@ -44,6 +47,7 @@
     "tsdown": "^0.13.0",
     "tsx": "^4.19.2",
     "typescript": "^5.8.2",
-    "vitest": "^3.2.4"
+    "vitest": "^3.2.4",
+    "wrangler": "^3.0.0"
   }
 }
