+++ apps/backend/src/routers/health.ts
@@ -1,0 +1,8 @@
+import { publicProcedure } from '../lib/orpc'
+
+export const healthRouter = {
+  // Health check
+  healthCheck: publicProcedure.handler(() => {
+    return 'OK'
+  }),
+}
