+++ apps/backend/src/routers/user.ts
@@ -1,0 +1,10 @@
+import { protectedProcedure } from '../lib/orpc'
+
+export const userRouter = {
+  // Get current user profile
+  profile: protectedProcedure.handler(({ context }) => {
+    return {
+      user: context.session?.user,
+    }
+  }),
+}
