+++ apps/backend/wrangler.toml
@@ -1,0 +1,33 @@
+name = "ai-meeting-digest-backend"
+main = "dist/worker.js"
+compatibility_date = "2024-01-01"
+compatibility_flags = ["nodejs_compat"]
+
+[build]
+command = "npm run build:worker"
+
+# Environment variables for Cloudflare Workers
+[vars]
+NODE_ENV = "production"
+
+# Secrets (set via wrangler secret put)
+# OPENROUTER_API_KEY = "your-openrouter-api-key"
+# DATABASE_URL = "your-database-url"
+# BETTER_AUTH_SECRET = "your-auth-secret"
+# CORS_ORIGINS = "https://your-frontend-domain.com"
+
+# D1 Database binding (if using Cloudflare D1)
+# [[d1_databases]]
+# binding = "DB"
+# database_name = "ai-meeting-digest"
+# database_id = "your-database-id"
+
+# KV namespace bindings (if needed for caching)
+# [[kv_namespaces]]
+# binding = "CACHE"
+# id = "your-kv-namespace-id"
+
+# R2 bucket bindings (if using for file storage)
+# [[r2_buckets]]
+# binding = "FILES"
+# bucket_name = "ai-meeting-digest-files"
